add font, edit login page, router, add wrapper and image for login
