+++ src/Pages/Login/LineBreak.jsx
@@ -1,0 +1,14 @@
+const LineBreak = () => {
+    return (
+        <div className="d-flex justify-content-between">
+            <div className="line-custom">
+                <hr />
+            </div>
+            <span className="opacity-50">Or</span>
+            <div className="line-custom">
+                <hr />
+            </div>
+        </div>
+    );
+};
+export default LineBreak;
