+++ src/index.css
@@ -2,4 +2,6 @@
   margin: 0;
   padding: 0;
   box-sizing: border-box;
+  font-family:'Montserrat', sans-serif;
+
 }
